+++ CHANGELOG.md
@@ -1,14 +1,23 @@
 # Change Log
-- 0.0.4
+# 0.0.5
+
+> #### Added
+> - alias autocomplete
+> - keyword autocomplete 
+> - lambeu malformated value error
+> - camel case warning and autocomplete
+
+# 0.0.4
 
 #### Added an icon to the extension 
 
 #### Added penetrou to the type keywords
 
-- Initial release
+# Initial release
 
->  #### Added syntax highlights
-> - Added snippets:
+#### Added syntax highlights
+
+> #### Added snippets:
 > - add
 > - sub
 > - push